refactor(pickers): split pr and issue surfaces (#528)
Problem: pickers.lua still contained the full PR and issue picker
surfaces, including their cache patching, local mutations, and action
wiring. That kept the largest picker workflows coupled to the remaining
checks, CI, release, and git surfaces.

Solution: move the PR picker flow into forge.picker.pr and the issue
picker flow into forge.picker.issue, then keep pickers.lua as a thin
public delegator for those surfaces.
