+++ lua/forge/pickers.lua
@@ -1,6 +1,5 @@
 local M = {}
 
-local availability = require('forge.availability')
 local ci = require('forge.ci')
 local log = require('forge.logger')
 local ops = require('forge.ops')
@@ -682,947 +681,14 @@ end
 ---@param f forge.Forge
 ---@param opts? forge.PickerLimitOpts
 function M.pr(state, f, opts)
-  opts = opts or {}
-  local next_state = ({ all = 'open', open = 'closed', closed = 'all' })[state]
-  local state_label = ({ all = 'All', open = 'Open', closed = 'Closed' })[state] or state
-  local forge_mod = require('forge')
-  local cfg = forge_mod.config()
-  local limits = limit_settings(cfg.display.limits.pulls, opts.limit)
-  local limit_step = limits.step
-  local visible_limit = limits.visible
-  local fetch_limit = limits.fetch
-  local use_cache = limits.use_cache
-  local ref = scoped_forge_ref(f, opts.scope)
-  local scope_suffix = scoped_key(forge_mod, ref)
-  local cache_key = scoped_list_key('pr', state, scope_suffix)
-  local pr_fields = f.pr_fields
-  local num_field = pr_fields.number
-  local pr_state_field = pr_fields.state
-  local show_state = state ~= 'open'
-  local current_limit = visible_limit
-  local current_prs
-  local prs_stale = true
-  local picker_handle
-
-  local function build_pr_entries(prs, limit)
-    limit = limit or current_limit
-
-    table.sort(prs, function(a, b)
-      return (a[num_field] or 0) > (b[num_field] or 0)
-    end)
-    local has_more = #prs > limit
-    if has_more then
-      prs = vim.list_slice(prs, 1, limit)
-    end
-    local entries = {}
-    local rows_for = cached_rows(function(width)
-      return forge_mod.format_prs(prs, pr_fields, show_state, { width = width })
-    end)
-    local displays = rows_for()
-    for i, pr in ipairs(prs) do
-      local num = tostring(pr[pr_fields.number] or '')
-      local draft_field = rawget(pr_fields, 'is_draft')
-      table.insert(entries, {
-        display = displays[i],
-        render_display = function(width)
-          return rows_for(width)[i]
-        end,
-        value = {
-          num = num,
-          scope = ref,
-          state = pr[pr_fields.state],
-          is_draft = draft_field and pr[draft_field] or nil,
-        },
-        ordinal = (pr[pr_fields.title] or '') .. ' #' .. num,
-      })
-    end
-    local count = #entries
-    if has_more then
-      entries[#entries + 1] = load_more_entry(expanded_limit(limit, limit_step), true)
-    end
-    local empty_text = state == 'all' and ('No %s'):format(f.labels.pr)
-      or ('No %s %s'):format(state, f.labels.pr)
-    return with_placeholder(entries, empty_text), count
-  end
-
-  ---@param emit fun(entry: forge.PickerEntry?)
-  local function emit_cached_prs(emit)
-    local entries = build_pr_entries(current_prs, current_limit)
-    for _, entry in ipairs(entries) do
-      emit(entry)
-    end
-    emit(nil)
-  end
-
-  local function maybe_prefetch_next()
-    if not use_cache or not f.list_pr_json_cmd then
-      return
-    end
-    maybe_prefetch_list(
-      'pr',
-      next_state,
-      f.labels.pr,
-      f:list_pr_json_cmd(next_state, fetch_limit, ref),
-      scope_suffix
-    )
-  end
-
-  ---@param emit fun(entry: forge.PickerEntry?)
-  local function stream_prs(emit)
-    if current_prs and not prs_stale then
-      emit_cached_prs(emit)
-      return
-    end
-    log.debug(('fetching %s list (%s)...'):format(f.labels.pr, state))
-    picker_session.request_json(
-      cache_key,
-      f:list_pr_json_cmd(state, current_limit + 1, ref),
-      function(ok, prs, failure, stale)
-        if stale then
-          emit(nil)
-          return
-        end
-        if not ok then
-          log.error(picker_failure_text(failure, 'failed to fetch ' .. f.labels.pr))
-          emit(picker_failure_entry(failure, 'Failed to fetch ' .. f.labels.pr))
-          emit(nil)
-          return
-        end
-        current_prs = prs
-        prs_stale = false
-        if use_cache then
-          state_mod.set_list(cache_key, prs)
-        end
-        emit_cached_prs(emit)
-        maybe_prefetch_next()
-      end
-    )
-  end
-
-  local function rerender_pr_list()
-    if refresh_picker(picker_handle) then
-      return
-    end
-    M.pr(state, f, { limit = current_limit, back = opts.back, scope = ref })
-  end
-
-  local function refresh_pr_list()
-    prs_stale = true
-    rerender_pr_list()
-  end
-
-  local function reopen_list()
-    clear_state_caches('pr', scope_suffix)
-    state_mod.clear_pr_state(nil, ref)
-    refresh_pr_list()
-  end
-
-  local function pr_cache_key(list_state)
-    return scoped_list_key('pr', list_state, scope_suffix)
-  end
-
-  local function observed_open_pr_state(rows)
-    if type(rows) ~= 'table' then
-      return nil
-    end
-    for _, pr in ipairs(rows) do
-      local value = pr[pr_state_field]
-      if type(value) == 'string' then
-        local text = vim.trim(value)
-        local lower = text:lower()
-        if lower == 'open' or lower == 'opened' then
-          return text
-        end
-      end
-    end
-    return nil
-  end
-
-  local function observed_merged_pr_state(rows)
-    if type(rows) ~= 'table' then
-      return nil
-    end
-    for _, pr in ipairs(rows) do
-      local value = pr[pr_state_field]
-      if type(value) == 'string' then
-        local text = vim.trim(value)
-        if text:lower() == 'merged' then
-          return text
-        end
-      end
-    end
-    return nil
-  end
-
-  local function next_pr_state(current_pr, verb)
-    local current_state = current_pr[pr_state_field]
-    local current_text = type(current_state) == 'string' and vim.trim(current_state) or ''
-    if verb == 'close' then
-      if current_text ~= '' and current_text == current_text:upper() then
-        return 'CLOSED'
-      end
-      return 'closed'
-    end
-    local open_state = observed_open_pr_state(state == 'open' and current_prs or nil)
-      or observed_open_pr_state(state == 'all' and current_prs or nil)
-      or observed_open_pr_state(state_mod.get_list(pr_cache_key('open')))
-      or observed_open_pr_state(state_mod.get_list(pr_cache_key('all')))
-    if open_state then
-      return open_state
-    end
-    if current_text ~= '' and current_text == current_text:upper() then
-      return 'OPEN'
-    end
-    return 'open'
-  end
-
-  local function merged_pr_state(current_pr)
-    local merged_state = observed_merged_pr_state(state == 'all' and current_prs or nil)
-      or observed_merged_pr_state(state_mod.get_list(pr_cache_key('all')))
-    if merged_state then
-      return merged_state
-    end
-    local current_state = current_pr[pr_state_field]
-    local current_text = type(current_state) == 'string' and vim.trim(current_state) or ''
-    if current_text ~= '' and current_text == current_text:upper() then
-      return 'MERGED'
-    end
-    return 'merged'
-  end
-
-  local function patch_pr_cache(list_state, mutate)
-    local key = pr_cache_key(list_state)
-    local prs = state_mod.get_list(key)
-    if type(prs) ~= 'table' then
-      return
-    end
-    mutate(prs)
-    state_mod.set_list(key, prs)
-  end
-
-  local function revalidate_current_prs()
-    picker_session.request_json(
-      cache_key,
-      f:list_pr_json_cmd(state, current_limit + 1, ref),
-      function(ok, prs, failure, stale)
-        if stale then
-          return
-        end
-        if not ok then
-          log.error(picker_failure_text(failure, 'failed to fetch ' .. f.labels.pr))
-          return
-        end
-        current_prs = prs
-        prs_stale = false
-        if use_cache then
-          state_mod.set_list(cache_key, prs)
-        end
-        rerender_pr_list()
-        maybe_prefetch_next()
-      end
-    )
-  end
-
-  ---@param entry forge.PickerEntry
-  local function locally_approve_pr(entry)
-    local scope = entry.value.scope or ref
-    ---@type forge.PRState
-    local current_pr_state = vim.tbl_extend('force', {
-      state = entry.value.state or 'OPEN',
-      mergeable = 'UNKNOWN',
-      review_decision = '',
-      is_draft = entry.value.is_draft == true,
-    }, vim.deepcopy(state_mod.pr_state(f, entry.value.num, scope) or {}))
-    current_pr_state.review_decision = 'APPROVED'
-    state_mod.set_pr_state(entry.value.num, current_pr_state, scope)
-    rerender_pr_list()
-    revalidate_current_prs()
-  end
-
-  ---@param entry forge.PickerEntry
-  local function locally_merge_pr(entry)
-    local current_index, current_pr = list_row(current_prs, num_field, entry.value.num)
-    if not current_index or type(current_pr) ~= 'table' then
-      refresh_pr_list()
-      return
-    end
-    local scope = entry.value.scope or ref
-    if state == 'all' then
-      local updated_pr = vim.deepcopy(current_pr)
-      updated_pr[pr_state_field] = merged_pr_state(current_pr)
-      current_prs[current_index] = updated_pr
-    else
-      table.remove(current_prs, current_index)
-    end
-    if use_cache then
-      state_mod.set_list(cache_key, current_prs)
-    end
-
-    ---@type forge.PRState
-    local current_pr_state = vim.tbl_extend('force', {
-      state = entry.value.state or 'OPEN',
-      mergeable = 'UNKNOWN',
-      review_decision = '',
-      is_draft = entry.value.is_draft == true,
-    }, vim.deepcopy(state_mod.pr_state(f, entry.value.num, scope) or {}))
-    current_pr_state.state = merged_pr_state(current_pr)
-    current_pr_state.is_draft = false
-    state_mod.set_pr_state(entry.value.num, current_pr_state, scope)
-
-    for _, list_state in ipairs(list_states) do
-      if list_state ~= state then
-        clear_list_cache(pr_cache_key(list_state))
-      end
-    end
-
-    rerender_pr_list()
-    revalidate_current_prs()
-  end
-
-  ---@param entry forge.PickerEntry
-  local function locally_toggle_pr_draft(entry)
-    local scope = entry.value.scope or ref
-    ---@type forge.PRState
-    local current_pr_state = vim.tbl_extend('force', {
-      state = entry.value.state or 'OPEN',
-      mergeable = 'UNKNOWN',
-      review_decision = '',
-      is_draft = entry.value.is_draft == true,
-    }, vim.deepcopy(state_mod.pr_state(f, entry.value.num, scope) or {}))
-    current_pr_state.is_draft = not current_pr_state.is_draft
-    state_mod.set_pr_state(entry.value.num, current_pr_state, scope)
-    rerender_pr_list()
-    revalidate_current_prs()
-  end
-
-  local function locally_toggle_pr(entry, verb)
-    local current_index, current_pr = list_row(current_prs, num_field, entry.value.num)
-    if not current_index or type(current_pr) ~= 'table' then
-      refresh_pr_list()
-      return
-    end
-    local updated_pr = vim.deepcopy(current_pr)
-    updated_pr[pr_state_field] = next_pr_state(current_pr, verb)
-
-    if state == 'all' then
-      current_prs[current_index] = updated_pr
-    else
-      table.remove(current_prs, current_index)
-    end
-    if use_cache then
-      state_mod.set_list(cache_key, current_prs)
-    end
-
-    if verb == 'close' then
-      if state ~= 'closed' then
-        patch_pr_cache('closed', function(prs)
-          upsert_list_row(prs, num_field, entry.value.num, vim.deepcopy(updated_pr))
-        end)
-      end
-      if state ~= 'open' then
-        patch_pr_cache('open', function(prs)
-          remove_list_row(prs, num_field, entry.value.num)
-        end)
-      end
-    else
-      if state ~= 'open' then
-        patch_pr_cache('open', function(prs)
-          upsert_list_row(prs, num_field, entry.value.num, vim.deepcopy(updated_pr))
-        end)
-      end
-      if state ~= 'closed' then
-        patch_pr_cache('closed', function(prs)
-          remove_list_row(prs, num_field, entry.value.num)
-        end)
-      end
-    end
-    if state ~= 'all' then
-      patch_pr_cache('all', function(prs)
-        upsert_list_row(prs, num_field, entry.value.num, vim.deepcopy(updated_pr))
-      end)
-    end
-
-    rerender_pr_list()
-    revalidate_current_prs()
-  end
-
-  local function back_to_list()
-    M.pr(state, f, { limit = current_limit, back = opts.back, scope = ref })
-  end
-
-  local function pr_entity_only(entry)
-    return entity_row(entry)
-  end
-
-  local function pr_load_more_or_entity(entry)
-    return entity_row(entry) or load_more_row(entry)
-  end
-
-  local function pr_create_visible(entry)
-    local kind = picker_row_kind(entry)
-    return kind == 'none'
-      or kind == 'load_more'
-      or kind == 'empty'
-      or kind == 'error'
-      or kind == 'entity'
-  end
-
-  local function pr_filter_visible(entry)
-    local kind = picker_row_kind(entry)
-    return kind ~= 'error'
-  end
-
-  local function pr_refresh_visible(_)
-    return true
-  end
-
-  local actions = {
-    {
-      name = 'default',
-      label = function(entry)
-        if load_more_row(entry) then
-          return 'load more'
-        end
-        return require('forge.review').label(f, entry and entry.value or nil)
-      end,
-      available = pr_load_more_or_entity,
-      fn = function(entry)
-        if entry and entry.load_more then
-          current_limit = entry.next_limit
-          prs_stale = true
-        elseif entry then
-          pr_action_fns(f, entry.value).review()
-        end
-      end,
-    },
-    {
-      name = 'ci',
-      label = 'checks',
-      reload = false,
-      available = pr_entity_only,
-      fn = function(entry)
-        if entry and not entry.load_more then
-          pr_action_fns(f, entry.value).ci({ back = back_to_list })
-        end
-      end,
-    },
-    {
-      name = 'edit',
-      label = 'edit',
-      available = pr_entity_only,
-      fn = function(entry)
-        if entry and not entry.load_more then
-          pr_action_fns(f, entry.value).edit()
-        end
-      end,
-    },
-    {
-      name = 'approve',
-      label = 'approve',
-      available = function(entry)
-        return availability.pr_can_approve(f, entry)
-      end,
-      fn = function(entry)
-        if entry and not entry.load_more then
-          ops.pr_approve(f, entry.value, {
-            on_success = function()
-              locally_approve_pr(entry)
-            end,
-            on_failure = reopen_list,
-          })
-        end
-      end,
-    },
-    {
-      name = 'merge',
-      label = 'merge',
-      available = function(entry)
-        return availability.pr_can_merge(f, entry)
-      end,
-      fn = function(entry)
-        if entry and not entry.load_more then
-          ops.pr_merge(f, entry.value, nil, {
-            on_success = function()
-              locally_merge_pr(entry)
-            end,
-            on_failure = reopen_list,
-          })
-        end
-      end,
-    },
-    {
-      name = 'create',
-      label = 'create',
-      available = pr_create_visible,
-      fn = function()
-        forge_mod.create_pr({ back = opts.back, scope = ref })
-      end,
-    },
-    {
-      name = 'toggle',
-      available = pr_toggle_entry,
-      label = function(entry)
-        if entry == nil then
-          return 'close/reopen'
-        end
-        return surface_policy.pr_toggle_verb(entry)
-      end,
-      fn = function(entry)
-        if not entry or entry.load_more then
-          return
-        end
-        local verb = surface_policy.pr_toggle_verb(entry)
-        local callbacks = {
-          on_success = function()
-            locally_toggle_pr(entry, verb)
-          end,
-          on_failure = refresh_pr_list,
-        }
-        if verb == 'close' then
-          ops.pr_close(f, entry.value, callbacks)
-        elseif verb == 'reopen' then
-          ops.pr_reopen(f, entry.value, callbacks)
-        end
-      end,
-    },
-    {
-      name = 'draft',
-      label = function(entry)
-        return availability.pr_draft_label(f, entry)
-      end,
-      available = function(entry)
-        return availability.pr_can_toggle_draft(f, entry)
-      end,
-      fn = function(entry)
-        if entry and not entry.load_more and f.capabilities.draft then
-          pr_toggle_draft_action(f, entry.value, {
-            on_success = function()
-              locally_toggle_pr_draft(entry)
-            end,
-            on_failure = reopen_list,
-          })
-        end
-      end,
-    },
-    {
-      name = 'filter',
-      label = 'filter',
-      reload = false,
-      available = pr_filter_visible,
-      fn = function()
-        M.pr(next_state, f, { limit = current_limit, back = opts.back, scope = ref })
-      end,
-    },
-    {
-      name = 'refresh',
-      label = 'refresh',
-      reload = false,
-      available = pr_refresh_visible,
-      fn = function()
-        clear_state_caches('pr', scope_suffix)
-        state_mod.clear_pr_state(nil, ref)
-        refresh_pr_list()
-      end,
-    },
-  }
-
-  local cached = use_cache and state_mod.get_list(cache_key) or nil
-  if cached then
-    current_prs = cached
-    prs_stale = false
-  end
-
-  local initial_prompt
-  if current_prs then
-    local _, count = build_pr_entries(current_prs, current_limit)
-    initial_prompt = ('%s %s (%d)> '):format(state_label, f.labels.pr, count)
-  else
-    initial_prompt = ('%s %s> '):format(state_label, f.labels.pr)
-  end
-
-  picker_handle = picker.pick({
-    prompt = initial_prompt,
-    entries = {},
-    actions = actions,
-    header_order = pr_header_order,
-    picker_name = 'pr',
-    back = opts.back,
-    stream = stream_prs,
-  })
-
-  if current_prs then
-    maybe_prefetch_next()
-  end
+  require('forge.picker.pr').pick(state, f, opts)
 end
 
 ---@param state 'all'|'open'|'closed'
 ---@param f forge.Forge
 ---@param opts? forge.PickerLimitOpts
 function M.issue(state, f, opts)
-  opts = opts or {}
-  local next_state = ({ all = 'open', open = 'closed', closed = 'all' })[state]
-  local state_label = ({ all = 'All', open = 'Open', closed = 'Closed' })[state] or state
-  local forge_mod = require('forge')
-  local cfg = forge_mod.config()
-  local limits = limit_settings(cfg.display.limits.issues, opts.limit)
-  local limit_step = limits.step
-  local visible_limit = limits.visible
-  local fetch_limit = limits.fetch
-  local use_cache = limits.use_cache
-  local ref = scoped_forge_ref(f, opts.scope)
-  local scope_suffix = scoped_key(forge_mod, ref)
-  local cache_key = scoped_list_key('issue', state, scope_suffix)
-  local issue_fields = f.issue_fields
-  local num_field = issue_fields.number
-  local issue_state_field = issue_fields.state
-  local issue_show_state = state == 'all'
-  local current_limit = visible_limit
-  local current_issues
-  local issues_stale = true
-  local picker_handle
-
-  local function build_issue_entries(issues, limit)
-    limit = limit or current_limit
-
-    table.sort(issues, function(a, b)
-      return (a[num_field] or 0) > (b[num_field] or 0)
-    end)
-    local has_more = #issues > limit
-    if has_more then
-      issues = vim.list_slice(issues, 1, limit)
-    end
-    local state_field = issue_fields.state
-    local entries = {}
-    local rows_for = cached_rows(function(width)
-      return forge_mod.format_issues(issues, issue_fields, issue_show_state, { width = width })
-    end)
-    local displays = rows_for()
-    for i, issue in ipairs(issues) do
-      local n = tostring(issue[num_field] or '')
-      table.insert(entries, {
-        display = displays[i],
-        render_display = function(width)
-          return rows_for(width)[i]
-        end,
-        value = { num = n, scope = ref, state = issue[state_field] },
-        ordinal = (issue[issue_fields.title] or '') .. ' #' .. n,
-      })
-    end
-    local count = #entries
-    if has_more then
-      entries[#entries + 1] = load_more_entry(expanded_limit(limit, limit_step), true)
-    end
-    local empty_text = state == 'all' and ('No %s'):format(f.labels.issue)
-      or ('No %s %s'):format(state, f.labels.issue)
-    return with_placeholder(entries, empty_text), count
-  end
-
-  ---@param emit fun(entry: forge.PickerEntry?)
-  local function emit_cached_issues(emit)
-    local entries = build_issue_entries(current_issues, current_limit)
-    for _, entry in ipairs(entries) do
-      emit(entry)
-    end
-    emit(nil)
-  end
-
-  local function maybe_prefetch_next()
-    if not use_cache or not f.list_issue_json_cmd then
-      return
-    end
-    maybe_prefetch_list(
-      'issue',
-      next_state,
-      f.labels.issue,
-      f:list_issue_json_cmd(next_state, fetch_limit, ref),
-      scope_suffix
-    )
-  end
-
-  ---@param emit fun(entry: forge.PickerEntry?)
-  local function stream_issues(emit)
-    if current_issues and not issues_stale then
-      emit_cached_issues(emit)
-      return
-    end
-    log.debug('fetching issue list (' .. state .. ')...')
-    picker_session.request_json(
-      cache_key,
-      f:list_issue_json_cmd(state, current_limit + 1, ref),
-      function(ok, issues, failure, stale)
-        if stale then
-          emit(nil)
-          return
-        end
-        if not ok then
-          log.error(picker_failure_text(failure, 'failed to fetch issues'))
-          emit(picker_failure_entry(failure, 'Failed to fetch issues'))
-          emit(nil)
-          return
-        end
-        current_issues = issues
-        issues_stale = false
-        if use_cache then
-          state_mod.set_list(cache_key, issues)
-        end
-        emit_cached_issues(emit)
-        maybe_prefetch_next()
-      end
-    )
-  end
-
-  local function rerender_issue_list()
-    if refresh_picker(picker_handle) then
-      return
-    end
-    M.issue(state, f, { limit = current_limit, back = opts.back, scope = ref })
-  end
-
-  local function refresh_issue_list()
-    issues_stale = true
-    rerender_issue_list()
-  end
-
-  local function issue_cache_key(list_state)
-    return scoped_list_key('issue', list_state, scope_suffix)
-  end
-
-  local function observed_open_issue_state(rows)
-    if type(rows) ~= 'table' then
-      return nil
-    end
-    for _, issue in ipairs(rows) do
-      local value = issue[issue_state_field]
-      if type(value) == 'string' then
-        local text = vim.trim(value)
-        local lower = text:lower()
-        if lower == 'open' or lower == 'opened' then
-          return text
-        end
-      end
-    end
-    return nil
-  end
-
-  local function next_issue_state(current_issue, verb)
-    local current_state = current_issue[issue_state_field]
-    local current_text = type(current_state) == 'string' and vim.trim(current_state) or ''
-    if verb == 'close' then
-      if current_text ~= '' and current_text == current_text:upper() then
-        return 'CLOSED'
-      end
-      return 'closed'
-    end
-    local open_state = observed_open_issue_state(state == 'open' and current_issues or nil)
-      or observed_open_issue_state(state == 'all' and current_issues or nil)
-      or observed_open_issue_state(state_mod.get_list(issue_cache_key('open')))
-      or observed_open_issue_state(state_mod.get_list(issue_cache_key('all')))
-    if open_state then
-      return open_state
-    end
-    if current_text ~= '' and current_text == current_text:upper() then
-      return 'OPEN'
-    end
-    return 'open'
-  end
-
-  local function patch_issue_cache(list_state, mutate)
-    local key = issue_cache_key(list_state)
-    local issues = state_mod.get_list(key)
-    if type(issues) ~= 'table' then
-      return
-    end
-    mutate(issues)
-    state_mod.set_list(key, issues)
-  end
-
-  local function revalidate_current_issues()
-    picker_session.request_json(
-      cache_key,
-      f:list_issue_json_cmd(state, current_limit + 1, ref),
-      function(ok, issues, failure, stale)
-        if stale then
-          return
-        end
-        if not ok then
-          log.error(picker_failure_text(failure, 'failed to fetch issues'))
-          return
-        end
-        current_issues = issues
-        issues_stale = false
-        if use_cache then
-          state_mod.set_list(cache_key, issues)
-        end
-        rerender_issue_list()
-        maybe_prefetch_next()
-      end
-    )
-  end
-
-  local function locally_toggle_issue(entry, verb)
-    local current_index, current_issue = list_row(current_issues, num_field, entry.value.num)
-    if not current_index or type(current_issue) ~= 'table' then
-      refresh_issue_list()
-      return
-    end
-    local updated_issue = vim.deepcopy(current_issue)
-    updated_issue[issue_state_field] = next_issue_state(current_issue, verb)
-
-    if state == 'all' then
-      current_issues[current_index] = updated_issue
-    else
-      table.remove(current_issues, current_index)
-    end
-    if use_cache then
-      state_mod.set_list(cache_key, current_issues)
-    end
-
-    if verb == 'close' then
-      if state ~= 'closed' then
-        patch_issue_cache('closed', function(issues)
-          upsert_list_row(issues, num_field, entry.value.num, vim.deepcopy(updated_issue))
-        end)
-      end
-      if state ~= 'open' then
-        patch_issue_cache('open', function(issues)
-          remove_list_row(issues, num_field, entry.value.num)
-        end)
-      end
-    else
-      if state ~= 'open' then
-        patch_issue_cache('open', function(issues)
-          upsert_list_row(issues, num_field, entry.value.num, vim.deepcopy(updated_issue))
-        end)
-      end
-      if state ~= 'closed' then
-        patch_issue_cache('closed', function(issues)
-          remove_list_row(issues, num_field, entry.value.num)
-        end)
-      end
-    end
-    if state ~= 'all' then
-      patch_issue_cache('all', function(issues)
-        upsert_list_row(issues, num_field, entry.value.num, vim.deepcopy(updated_issue))
-      end)
-    end
-
-    rerender_issue_list()
-    revalidate_current_issues()
-  end
-
-  local actions = {
-    {
-      name = 'default',
-      label = 'open',
-      close = false,
-      fn = function(entry)
-        if entry and entry.load_more then
-          current_limit = entry.next_limit
-          issues_stale = true
-        elseif entry then
-          issue_action_fns(f, entry.value).browse()
-        end
-      end,
-    },
-    {
-      name = 'browse',
-      label = 'web',
-      close = false,
-      fn = function(entry)
-        if entry and not entry.load_more then
-          issue_action_fns(f, entry.value).browse()
-        end
-      end,
-    },
-    {
-      name = 'edit',
-      label = 'edit',
-      fn = function(entry)
-        if entry and not entry.load_more then
-          issue_action_fns(f, entry.value).edit()
-        end
-      end,
-    },
-    {
-      name = 'toggle',
-      available = issue_toggle_entry,
-      label = function(entry)
-        if entry == nil then
-          return 'close/reopen'
-        end
-        return surface_policy.issue_toggle_verb(entry)
-      end,
-      fn = function(entry)
-        if not entry or entry.load_more then
-          return
-        end
-        local verb = surface_policy.issue_toggle_verb(entry)
-        local callbacks = {
-          on_success = function()
-            locally_toggle_issue(entry, verb)
-          end,
-          on_failure = refresh_issue_list,
-        }
-        if verb == 'close' then
-          ops.issue_close(f, entry.value, callbacks)
-        elseif verb == 'reopen' then
-          ops.issue_reopen(f, entry.value, callbacks)
-        end
-      end,
-    },
-    {
-      name = 'create',
-      label = 'create',
-      fn = function()
-        forge_mod.create_issue({ back = opts.back, scope = ref })
-      end,
-    },
-    {
-      name = 'filter',
-      label = 'filter',
-      reload = false,
-      fn = function()
-        M.issue(next_state, f, { limit = current_limit, back = opts.back, scope = ref })
-      end,
-    },
-    {
-      name = 'refresh',
-      label = 'refresh',
-      reload = false,
-      fn = function()
-        clear_state_caches('issue', scope_suffix)
-        refresh_issue_list()
-      end,
-    },
-  }
-
-  local cached = use_cache and state_mod.get_list(cache_key) or nil
-  if cached then
-    current_issues = cached
-    issues_stale = false
-  end
-
-  local initial_prompt
-  if current_issues then
-    local _, count = build_issue_entries(current_issues, current_limit)
-    initial_prompt = ('%s %s (%d)> '):format(state_label, f.labels.issue, count)
-  else
-    initial_prompt = ('%s %s> '):format(state_label, f.labels.issue)
-  end
-
-  picker_handle = picker.pick({
-    prompt = initial_prompt,
-    entries = {},
-    actions = actions,
-    header_order = issue_header_order,
-    picker_name = 'issue',
-    back = opts.back,
-    stream = stream_issues,
-  })
-
-  if current_issues then
-    maybe_prefetch_next()
-  end
+  require('forge.picker.issue').pick(state, f, opts)
 end
 
 ---@param f forge.Forge
