+++ spec/pickers_spec.lua
@@ -21,6 +21,8 @@ local preload_modules = {
   'forge.logger',
   'forge.ops',
   'forge.picker',
+  'forge.picker.issue',
+  'forge.picker.pr',
   'forge.picker.shared',
   'forge.surface_policy',
   'forge.state',
@@ -32,6 +34,8 @@ local loaded_modules = {
   'forge.logger',
   'forge.ops',
   'forge.picker',
+  'forge.picker.issue',
+  'forge.picker.pr',
   'forge.picker.shared',
   'forge.picker.session',
   'forge.pickers',
